+++ CHANGELOG.md
@@ -3,7 +3,9 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
-## [v6.0.0-1.19.4] - 2023-05-24
-- Ported to Minecraft 1.19.4
+## [v7.0.0-1.20] - 2023-06-10
+- Ported to Minecraft 1.20
+### Fixed
+- Row counts above 9 now render correctly
 
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/